Fix hydration warning on hero section of the home page as long as dynamically loading background animation
