+++ src/app/_components/homepage/HeroSection/DynamicAnimatedBackground.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+const DynamicAnimatedBackground = dynamic(
+  () => import("./AnimatedBackground"),
+  {
+    ssr: false,
+  },
+);
+
+export default DynamicAnimatedBackground;
