+++ src/app/changelog/page.tsx
@@ -40,7 +40,7 @@ export default function ChangelogPage() {
   };
 
   return (
-    <div suppressHydrationWarning>
+    <div>
       <script
         type="application/ld+json"
         dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}
